requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-login
flask-migrate
flask-Mail
flask-dotenv
python-dotenv
pyjwt